src/requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.6
Django==3.2.6
django-cors-headers==3.8.0
djangorestframework==3.12.4
idna==3.2
mysql-client==0.0.1
mysql-connector-python==8.0.26
mysqlclient==2.0.3
protobuf==3.17.3
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
gunicorn
